+++ app/services/baseService.js
@@ -3,8 +3,8 @@ class BaseService {
     this.model = model;
   }
 
-  async findAll({ page = 1, limit = 10, sortField = 'id', sortOrder = 'ASC', filter = {} } = {}) {
-    return this.model.findAll({ page, limit, sortField, sortOrder, filter });
+  async findAll({ page = 1, limit = 10, sortField = 'id', sortOrder = 'ASC', ...filter } = {}) {
+    return this.model.findAll({ page, limit, sortField, sortOrder, ...filter });
   }
 
   async findById(id) {