+++ CHANGELOG.md
@@ -6,11 +6,12 @@ All significant changes to this project are documented in this file.
 
 ### Refactored
 
-- Refactored and enhanced services for cities, departments, districts, and neighborhoods.
+- Refactored and enhanced services for cities, departments, districts, and neighborhoods
+- Refactored and enhanced controllers cities, departments, districts, and neighborhoods
 
 ### Documentation
 
-- Updated `setup.md` to correct the database `.tar.gz` filename.
+- Updated `setup.md` to correct the database `.tar.gz` filename
 
 ## [2.16.0] - 2025-05-06